fix: add route guard for merge page
