+++ src/app/routes/Router.tsx
@@ -26,7 +26,10 @@ import {
 } from '../../lib'
 import { OverviewSection } from '../../types'
 import { Layout, Breadcrumbs, BreadcrumbItem } from '../layout'
-import { CheckAuthorityForSection } from './CheckAuthorityForSection'
+import {
+    SectionAuthorizedGuard,
+    SectionCanMergeGuard,
+} from './AuthorizationGuards'
 import { DefaultErrorRoute } from './DefaultErrorRoute'
 import { LegacyAppRedirect } from './LegacyAppRedirect'
 import { RouteHandle } from './types'
@@ -170,22 +173,24 @@ const schemaSectionRoutes = Object.values(SCHEMA_SECTIONS).map((section) => (
                 />
             )}
             {isMergableSection(section) && (
-                <Route
-                    path={routePaths.merge}
-                    lazy={createSectionLazyRouteFunction(section, 'Merge')}
-                    handle={
-                        {
-                            crumb: (matchInfo) => (
-                                <BreadcrumbItem
-                                    label={i18n.t('Merge {{modelName}}', {
-                                        modelName: section.titlePlural,
-                                    })}
-                                    to={matchInfo.pathname}
-                                />
-                            ),
-                        } satisfies RouteHandle
-                    }
-                />
+                <Route element={<SectionCanMergeGuard />}>
+                    <Route
+                        path={routePaths.merge}
+                        lazy={createSectionLazyRouteFunction(section, 'Merge')}
+                        handle={
+                            {
+                                crumb: (matchInfo) => (
+                                    <BreadcrumbItem
+                                        label={i18n.t('Merge {{modelName}}', {
+                                            modelName: section.titlePlural,
+                                        })}
+                                        to={matchInfo.pathname}
+                                    />
+                                ),
+                            } satisfies RouteHandle
+                        }
+                    />
+                </Route>
             )}
             <Route path=":id" element={<VerifyModelId />}>
                 <Route
@@ -223,7 +228,7 @@ const routes = createRoutesFromElements(
             path="/"
             element={<Navigate to={routePaths.overviewRoot} replace />}
         />
-        <Route element={<CheckAuthorityForSection />}>
+        <Route element={<SectionAuthorizedGuard />}>
             <Route path={routePaths.overviewRoot}>
                 <Route
                     index
